Commit changes to design

+++ forms/doDeletePatient.php
@@ -2,6 +2,7 @@
 ob_start();
 session_start();
 include "../db_connect.php";
+include "../helper_functions.php";
 ?>
 
 <?php
@@ -12,6 +13,12 @@ if (!isset($_SESSION['admin_id'])) {
 
 if (isset($_POST['patient_id'])) {
     $id = $_POST['patient_id'];
+
+    if (check_patient_exist_by_payment_status($conn, $id) !== false) {
+        $_SESSION['message'] = "Patient record cannot be deleted as there are payment records associated with it.";
+        header("Location: patientDetails.php");
+        exit();
+    }
 
     $query = "DELETE FROM patient WHERE patient_id = ?";
     $stmt = mysqli_prepare($conn, $query);
@@ -31,7 +38,7 @@ if (isset($_POST['patient_id'])) {
         }
         mysqli_stmt_close($stmt);
     }
-} else{
+} else {
     header("Location: patientDetails.php");
     exit();
 }
